Added support for speakOnDemand(NVDA2024.X)
