+++ addon/globalPlugins/clock/__init__.py
@@ -7,11 +7,13 @@ from typing import Tuple
 import sys
 from . import skipTranslation
 import globalVars
+import versionInfo
 from . import alarmHandler
 import globalPluginHandler
 import gui
 from . import paths
 import scriptHandler
+import speech
 from logHandler import log
 import ui
 from . import formats
@@ -167,6 +169,35 @@ def checkLocalTimeFormats():
 			log.debugWarning(f"The format '{fmt}' appears {nbOccurrences} times for the current localization.")
 
 
+# supports speakOnDemand mode for NVDA2024.1 and later. Issue #52
+# Copyright (C) 2023-2024 Cyrille Bougot
+def isSpeechOnDemandFeatureAvailable():
+	"""Indicates if the speech on-demand feature is available in the current version of NVDA.
+	"""
+
+	try:
+		# NVDA >= 2024.1
+		speech.SpeechMode.onDemand
+		return True
+	except AttributeError:
+		# NVDA <= 2023.3
+		return False
+
+
+def getSpeechOnDemandParameter():
+	"""Returns a dictionary containing the speakOnDemand parameter and its True value for NVDA versions where
+	this feature exists; returns an empty dictionary otherwise.
+	Use `**getSpeechOnDemandParameter()` in script decorator's parameters to define `speakOnDemand=True` only
+	for NVDA versions where this feature is supported.
+	"""
+
+	if isSpeechOnDemandFeatureAvailable():
+		# Define on-demand parameter only if the feature is available.
+		return {'speakOnDemand': True}
+	else:
+		return {}
+
+
 class GlobalPlugin(globalPluginHandler.GlobalPlugin):
 
 	# Translators: Script category for Clock addon commands in input gestures dialog.
@@ -253,7 +284,8 @@ class GlobalPlugin(globalPluginHandler.GlobalPlugin):
 			"the current year and the days remaining before the end of the year."
 		),
 		category=globalCommands.SCRCAT_SYSTEM,
-		gesture="kb:NVDA+f12"
+		gesture="kb:NVDA+f12",
+		**getSpeechOnDemandParameter()
 	)
 	def script_reportTimeAndDate(self, gesture):
 		now = datetime.now()
